chore(shared-state): refactor to use query file
