+++ plugins/lime-plugin-mesh-wide-upgrade/src/meshUpgradeApi.tsx
@@ -1,4 +1,4 @@
-import { sharedStateQueries } from "components/shared-state/SharedStateTypes";
+import { sharedStateQueries } from "components/shared-state/SharedStateQueriesKeys";
 
 import {
     MeshWideRPCReturnTypes,
